refactor: remove circular dependency from `notifications.tsx` and `notifications-list.tsx`
