+++ development/circular-deps.jsonc
@@ -47,9 +47,5 @@
   [
     "ui/components/multichain/pages/send/components/index.ts",
     "ui/components/multichain/pages/send/components/your-accounts.tsx"
-  ],
-  [
-    "ui/pages/notifications/notifications-list.tsx",
-    "ui/pages/notifications/notifications.tsx"
   ]
 ]
