+++ ui/pages/notifications/notifications-list.tsx
@@ -14,7 +14,7 @@ import { useI18nContext } from '../../hooks/useI18nContext';
 import { NotificationsPlaceholder } from './notifications-list-placeholder';
 import { NotificationsListTurnOnNotifications } from './notifications-list-turn-on-notifications';
 import { NotificationsListItem } from './notifications-list-item';
-import { type Notification, TAB_KEYS } from './notifications';
+import type { Notification } from './notifications';
 import { NotificationsListReadAllButton } from './notifications-list-read-all-button';
 
 export type NotificationsListProps = {
@@ -24,6 +24,18 @@ export type NotificationsListProps = {
   isError: boolean;
   notificationsCount: number;
 };
+
+// NOTE - Tab filters could change once we support more notifications.
+export const enum TAB_KEYS {
+  // Shows all notifications
+  ALL = 'notifications-all-tab',
+
+  // These are only on-chain notifications (no snaps or feature announcements)
+  WALLET = 'notifications-wallet-tab',
+
+  // These are 3rd party notifications (snaps, feature announcements, web3 alerts)
+  WEB3 = 'notifications-other-tab',
+}
 
 function LoadingContent() {
   return (
